+++ src/mpd_client/search.h
@@ -19,6 +19,9 @@ bool mpd_client_search_add_to_plist_window(struct t_partition_state *partition_s
         const char *plist, unsigned to, const char *sort, bool sortdesc, unsigned start, unsigned end, sds *error);
 bool mpd_client_search_add_to_queue(struct t_partition_state *partition_state, const char *expression,
         unsigned to, enum mpd_position_whence whence, const char *sort, bool sortdesc, sds *error);
+bool mpd_client_search_add_to_queue_window(struct t_partition_state *partition_state, const char *expression,
+        unsigned to, enum mpd_position_whence whence, const char *sort, bool sortdesc,
+        unsigned start, unsigned end, sds *error);
 
 bool mpd_client_add_search_sort_param(struct t_partition_state *partition_state, const char *sort, bool sortdesc, bool check_version);
 bool mpd_client_add_search_group_param(struct mpd_connection *conn, enum mpd_tag_type tag);