+++ CHANGELOG.md
@@ -28,11 +28,14 @@ An another notable feature is the new list view that supplements the table and g
 - MYMPD_API_QUEUE_APPEND_PLAYLIST_RANGE: new
 - MYMPD_API_QUEUE_INSERT_PLAYLIST_RANGE: new
 - MYMPD_API_QUEUE_REPLACE_PLAYLIST_RANGE: new
+- MYMPD_API_QUEUE_APPEND_ALBUM_RANGE: new
+- MYMPD_API_QUEUE_INSERT_ALBUM_RANGE: new
+- MYMPD_API_QUEUE_REPLACE_ALBUM_RANGE: new
 - MYMPD_API_SETTINGS_GET: returns now available sticker types
 
 ### Changelog
 
-- Feat: Resume for songs
+- Feat: Resume for songs, playlists and albums #1338
 - Feat: Rating for albums and playlists #1134
 - Feat: User defined stickers #1091
 - Feat: Add list view